move AddCourseModal to seperate component
